+++ RepairSystem-Backend/src/main/java/ca/mcgill/ecse321/repairsystem/dto/AdministrativeAssistantDto.java
@@ -2,7 +2,7 @@ package ca.mcgill.ecse321.repairsystem.dto;
 
 public class AdministrativeAssistantDto extends PersonDto {
 	
-	public AdministrativeAssistantDto(String aName, int id, String aPassword, int aPhone, String aEmail) {
+	public AdministrativeAssistantDto(String aName, int id, String aPassword, long aPhone, String aEmail) {
 		super(aName,id, aPassword, aPhone, aEmail);
 	}
