+++ RepairSystem-Backend/src/main/java/ca/mcgill/ecse321/repairsystem/model/Mechanic.java
@@ -8,7 +8,7 @@ import javax.persistence.ManyToMany;
 public class Mechanic extends Person{
 
 
-	public Mechanic(String aName, int id, String aPassword, int aPhone, String aEmail, List<Service> allCapabilities)
+	public Mechanic(String aName, int id, String aPassword, long aPhone, String aEmail, List<Service> allCapabilities)
 	{
 	  super(aName, id, aPassword, aPhone, aEmail);    
 	  services = allCapabilities;
